nvme: move io-mgmt commands into a plugin
There are too many top-level commands cluttering the space. Move the
io-mgmt commands into a plugin, e.g.,

  nvme io-mgmt-recv -> nvme io-mgmt recv
